requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.7.0
arabic-reshaper==3.0.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.5
jose==1.0.0
MarkupSafe==3.0.2
motor==3.6.0
passlib==1.7.4
PyArabic==0.6.15
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.6.0
pymongo==4.9.2
python-bidi==0.6.3
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
rsa==4.9
secure-smtplib==0.1.1
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starkbank-ecdsa==2.2.0
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0